# Make Buf.send() faster, again