@chrisvest 2020-12-04
The `cleanerClose/heavy` case became slower, I think, because now it doesn't have much advantage compared to explicitly closed. Basically, those two cases aligned, while previously the cleaner case could upgrade the segment to a shared one and then forego its costs of closing.