@chrisvest 2020-12-04
I was worried that the transfer thread was perhaps not going to be a scalable solution on a loaded system, but it turns out that the performance benefit is even more pronounced then.

When running the benchmark with `-t 8`:

Before

```
Benchmark                    Mode  Cnt    Score   Error  Units
SendBenchmark.sendNonPooled  avgt   50  419,389 ± 6,732  us/op
SendBenchmark.sendPooled     avgt   50   17,409 ± 0,303  us/op
```

After

```
Benchmark                    Mode  Cnt   Score   Error  Units
SendBenchmark.sendNonPooled  avgt   50  42,229 ± 1,884  us/op
SendBenchmark.sendPooled     avgt   50  18,153 ± 0,347  us/op
```