@chrisvest 2020-12-04
@normanmaurer I'll split out the benchmark in a separate PR, since that'll be less controversial.