@chrisvest 2020-12-18
Closing this because it becomes a non-issue (no more thread-ownership transitions) when we remove thread-confinement in #21